requirements.txt
streamlit
pandas
numpy
scikit-learn
google-generativeai
python-dotenv
reportlab

